+++ src/shared/locales/pl-PL.ts
@@ -552,6 +552,8 @@ export const plPL = {
       "devtools": "Pokaż lub ukryj narzędzia deweloperskie",
       "nextTab": "Przejdź do następnej karty",
       "previousTab": "Przejdź do poprzedniej karty",
+      "directTab": "Przejdź do karty internetowej 1–8",
+      "lastTab": "Przejdź do ostatniej karty internetowej",
       "resetZoom": "Zresetuj powiększenie strony"
     },
     "responsive": {